requirements.txt
Flask==2.3.3
tensorflow>=2.16.0
numpy>=1.24.0
scikit-learn>=1.3.0
gunicorn>=21.0.0
Pillow>=10.0.0
matplotlib>=3.8.0
